Move common log and subprocess functions to a library
